+++ src/som/compiler/ast/method_generation_context.py
@@ -2,16 +2,16 @@ from collections import OrderedDict
 
 from rtruffle.source_section import SourceSection
 
-from .variable                                 import Argument, Local
+from .variable import Argument, Local
 
-from ..interpreter.nodes.field_node            import create_write_node, \
+from ...interpreter.nodes.field_node import create_write_node, \
                                                       create_read_node
-from ..interpreter.nodes.global_read_node      import \
+from ...interpreter.nodes.global_read_node import \
     UninitializedGlobalReadNode
-from ..interpreter.nodes.return_non_local_node import CatchNonLocalReturnNode
-from ..interpreter.invokable                   import Invokable
+from ...interpreter.nodes.return_non_local_node import CatchNonLocalReturnNode
+from ...interpreter.invokable import Invokable
 
-from ..vmobjects.primitive                     import empty_primitive
+from ...vmobjects.primitive import empty_primitive
 
 
 class MethodGenerationContext(object):
