+++ tests/basic_interpreter_test.py
@@ -1,7 +1,7 @@
 import unittest
 from parameterized import parameterized
 
-from som.compiler.parser import ParseError
+from som.compiler.ast.parser import ParseError
 
 from som.vm.universe       import Universe
